refactor: push list-of-buttons components to their own submodule
