+++ dev-client/src/screens/SettingsScreen/components/actions/LogOutButton.tsx
@@ -16,7 +16,7 @@
  */
 
 import {useTranslation} from 'react-i18next';
-import {ListButton} from 'terraso-mobile-client/components/buttons/ListButton';
+import {ListButton} from 'terraso-mobile-client/components/buttons/list/ListButton';
 import {LogoutModal} from 'terraso-mobile-client/components/modals/LogoutModal';
 
 export function LogOutButton() {
